server_decrypt/requirements.txt
phe
Flask-API
pymongo
dnspython
flask_cors
requests
pyopenssl